Update #6
Check to make sure that hide_set, hide_select, hide_viewport, hide_render is valid on both collection and objects before exporting.

Change how instance geoemetry is added to the list to avoid an issue where any object with digits at the end is considired valid.

Change how an object with zero materials is handled to avoid an error.